sdk fixes 1
