+++ bcccoreapi/client.go
@@ -1,4 +1,4 @@
-package coreapi
+package bcccoreapi
 
 import (
 	"fmt"
@@ -30,20 +30,20 @@ type Client struct {
 }
 
 const (
-	AffiliationsPath    = "/affiliations"
-	ConsentsPath        = "/consents"
-	CountriesPath       = "/countries"
-	GroupsPath          = "/groups"
-	OrgsPath            = "/v2/orgs"
-	PersonsPath         = "/v2/persons"
-	RelationsPath       = "/relations"
-	RoleAssignmentsPath = "/roleAssignments"
-	RolesPath           = "/roles"
+	affiliationsPath    = "/affiliations"
+	consentsPath        = "/consents"
+	countriesPath       = "/countries"
+	groupsPath          = "/groups"
+	orgsPath            = "/v2/orgs"
+	personsPath         = "/v2/persons"
+	relationsPath       = "/relations"
+	roleAssignmentsPath = "/roleAssignments"
+	rolesPath           = "/roles"
 )
 
 var DefaultAgent = fmt.Sprintf("Go-Coreapi/%s", Version)
 
-func New(url string, options ...ClientOption) *Client {
+func NewClient(url string, options ...ClientOption) *Client {
 	c := &Client{
 		http:      http.DefaultClient,
 		apiUrl:    url,
@@ -66,15 +66,15 @@ func New(url string, options ...ClientOption) *Client {
 		},
 	}
 
-	c.Affiliation = &genericClient[models.Affiliation, models.AffiliationWrite]{c, AffiliationsPath}
-	c.Consent = &genericClient[models.Consent, models.ConsentWrite]{c, ConsentsPath}
-	c.Country = &genericClient[models.Country, models.CountryWrite]{c, CountriesPath}
-	c.Group = &GroupClient{genericClient[models.Group, models.GroupWrite]{c, GroupsPath}}
-	c.Org = &genericClient[models.Org, models.OrgWrite]{c, OrgsPath}
-	c.Person = &genericClient[models.Person, models.PersonWrite]{c, PersonsPath}
-	c.Relation = &genericClient[models.Relation, models.RelationWrite]{c, RelationsPath}
-	c.RoleAssignment = &genericClient[models.RoleAssignment, models.RoleAssignmentWrite]{c, RoleAssignmentsPath}
-	c.Role = &genericClient[models.Role, models.RoleWrite]{c, RolesPath}
+	c.Affiliation = &genericClient[models.Affiliation, models.AffiliationWrite]{c, affiliationsPath}
+	c.Consent = &genericClient[models.Consent, models.ConsentWrite]{c, consentsPath}
+	c.Country = &genericClient[models.Country, models.CountryWrite]{c, countriesPath}
+	c.Group = &GroupClient{genericClient[models.Group, models.GroupWrite]{c, groupsPath}}
+	c.Org = &genericClient[models.Org, models.OrgWrite]{c, orgsPath}
+	c.Person = &genericClient[models.Person, models.PersonWrite]{c, personsPath}
+	c.Relation = &genericClient[models.Relation, models.RelationWrite]{c, relationsPath}
+	c.RoleAssignment = &genericClient[models.RoleAssignment, models.RoleAssignmentWrite]{c, roleAssignmentsPath}
+	c.Role = &genericClient[models.Role, models.RoleWrite]{c, rolesPath}
 
 	return c
 }
