+++ bcccoreapi/client_environment.go
@@ -1,4 +1,4 @@
-package coreapi
+package bcccoreapi
 
 type ClientCredentialsEnv struct {
 	TokenUrl string
