+++ bcccoreapi/client_request.go
@@ -1,4 +1,4 @@
-package coreapi
+package bcccoreapi
 
 import (
 	"bytes"
