+++ bcccoreapi/client_request_option.go
@@ -1,4 +1,4 @@
-package coreapi
+package bcccoreapi
 
 import (
 	"encoding/json"
