+++ bcccoreapi/client_request_option_test.go
@@ -1,4 +1,4 @@
-package coreapi
+package bcccoreapi
 
 import (
 	"fmt"
