+++ bcccoreapi/client_test.go
@@ -1,4 +1,4 @@
-package coreapi
+package bcccoreapi
 
 import (
 	"context"
@@ -16,7 +16,7 @@ func TestNewClient(t *testing.T) {
 }
 
 func getClientForTests(t *testing.T, scopes ...Scope) *Client {
-	c := New("http://localhost:3010", WithEmulator(context.Background(), ClientCredentialsEnv{
+	c := NewClient("http://localhost:3010", WithEmulator(context.Background(), ClientCredentialsEnv{
 		TokenUrl: "http://localhost:3020/token",
 		Audience: "localhost:3020",
 	}, scopes...))
