+++ bcccoreapi/consent_test.go
@@ -1,4 +1,4 @@
-package coreapi
+package bcccoreapi
 
 import (
 	"context"
@@ -11,41 +11,44 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestGetRoleAssignment(t *testing.T) {
-	c := getClientForTests(t, ScopePersonsRead, ScopePersonRoleAssignmentsRead)
+func TestGetConsent(t *testing.T) {
+	c := getClientForTests(t, ScopePersonsRead, ScopePersonConsentsRead)
 
-	res, err := c.RoleAssignment.Get(context.Background(), TestRoleAssignmentUID)
+	res, err := c.Consent.Get(context.Background(), TestConsentUID, Fields("*", "person.*"))
 
 	assert.NoError(t, err)
-	assert.Equal(t, TestRoleAssignmentUID, *res.Data.UID)
+	assert.Equal(t, TestConsentUID, *res.Data.UID)
+	assert.NotZero(t, TestPersonUID, res.Data.Person)
 }
 
-func TestFindRoleAssignment(t *testing.T) {
-	c := getClientForTests(t, ScopePersonRoleAssignmentsRead)
+func TestFindConsent(t *testing.T) {
+	c := getClientForTests(t, ScopePersonConsentsRead)
 
-	res, err := c.RoleAssignment.Find(context.Background(), Limit(2))
+	res, err := c.Consent.Find(context.Background(), Limit(2))
 
 	assert.NoError(t, err)
 	assert.Len(t, res.Data, 2)
 }
 
-func TestCreateAndUpdateRoleAssignment(t *testing.T) {
-	c := getClientForTests(t, ScopePersonRoleAssignmentsWrite)
+func TestCreateAndUpdateConsent(t *testing.T) {
+	c := getClientForTests(t, ScopePersonConsentsWrite)
 
-	body := models.RoleAssignmentWrite{
+	body := models.ConsentWrite{
+		OrgUID:    lo.ToPtr(TestOrgUID),
 		PersonUID: lo.ToPtr(TestPersonUID),
-		RoleUID:   lo.ToPtr(TestRoleUID),
+		Purpose:   models.ConsentPurposeDataSharing.Pointer(),
 		ValidFrom: lo.ToPtr(strfmt.DateTime(time.Now().Truncate(time.Millisecond))),
 	}
 
-	res, err := c.RoleAssignment.Create(context.Background(), body)
+	res, err := c.Consent.Create(context.Background(), body)
 	assert.NoError(t, err)
 
 	t.Run("CheckCreate", func(t *testing.T) {
 		assert.NotEmpty(t, res.Data.UID)
 
+		assert.Equal(t, body.OrgUID, res.Data.OrgUID)
 		assert.Equal(t, body.PersonUID, res.Data.PersonUID)
-		assert.Equal(t, body.RoleUID, res.Data.RoleUID)
+		assert.Equal(t, body.Purpose, res.Data.Purpose)
 		assert.True(t, body.ValidFrom.Equal(*res.Data.ValidFrom))
 		assert.Equal(t, body.ValidTo, res.Data.ValidTo)
 
@@ -53,13 +56,14 @@ func TestCreateAndUpdateRoleAssignment(t *testing.T) {
 
 	t.Run("CheckUpdate", func(t *testing.T) {
 		body.ValidTo = lo.ToPtr(strfmt.DateTime(time.Now().Truncate(time.Millisecond)))
-		res, err := c.RoleAssignment.Update(context.Background(), *res.Data.UID, body)
+		res, err := c.Consent.Update(context.Background(), *res.Data.UID, body)
 
 		assert.NoError(t, err)
 		assert.NotEmpty(t, res.Data.UID)
 
+		assert.Equal(t, body.OrgUID, res.Data.OrgUID)
 		assert.Equal(t, body.PersonUID, res.Data.PersonUID)
-		assert.Equal(t, body.RoleUID, res.Data.RoleUID)
+		assert.Equal(t, body.Purpose, res.Data.Purpose)
 		assert.True(t, body.ValidFrom.Equal(*res.Data.ValidFrom))
 		assert.True(t, body.ValidTo.Equal(*res.Data.ValidTo))
