+++ bcccoreapi/person_test.go
@@ -1,4 +1,4 @@
-package coreapi
+package bcccoreapi
 
 import (
 	"context"
