+++ bcccoreapi/role_assignment_test.go
@@ -1,4 +1,4 @@
-package coreapi
+package bcccoreapi
 
 import (
 	"context"
@@ -11,44 +11,41 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestGetConsent(t *testing.T) {
-	c := getClientForTests(t, ScopePersonsRead, ScopePersonConsentsRead)
+func TestGetRoleAssignment(t *testing.T) {
+	c := getClientForTests(t, ScopePersonsRead, ScopePersonRoleAssignmentsRead)
 
-	res, err := c.Consent.Get(context.Background(), TestConsentUID, Fields("*", "person.*"))
+	res, err := c.RoleAssignment.Get(context.Background(), TestRoleAssignmentUID)
 
 	assert.NoError(t, err)
-	assert.Equal(t, TestConsentUID, *res.Data.UID)
-	assert.NotZero(t, TestPersonUID, res.Data.Person)
+	assert.Equal(t, TestRoleAssignmentUID, *res.Data.UID)
 }
 
-func TestFindConsent(t *testing.T) {
-	c := getClientForTests(t, ScopePersonConsentsRead)
+func TestFindRoleAssignment(t *testing.T) {
+	c := getClientForTests(t, ScopePersonRoleAssignmentsRead)
 
-	res, err := c.Consent.Find(context.Background(), Limit(2))
+	res, err := c.RoleAssignment.Find(context.Background(), Limit(2))
 
 	assert.NoError(t, err)
 	assert.Len(t, res.Data, 2)
 }
 
-func TestCreateAndUpdateConsent(t *testing.T) {
-	c := getClientForTests(t, ScopePersonConsentsWrite)
+func TestCreateAndUpdateRoleAssignment(t *testing.T) {
+	c := getClientForTests(t, ScopePersonRoleAssignmentsWrite)
 
-	body := models.ConsentWrite{
-		OrgUID:    lo.ToPtr(TestOrgUID),
+	body := models.RoleAssignmentWrite{
 		PersonUID: lo.ToPtr(TestPersonUID),
-		Purpose:   models.ConsentPurposeDataSharing.Pointer(),
+		RoleUID:   lo.ToPtr(TestRoleUID),
 		ValidFrom: lo.ToPtr(strfmt.DateTime(time.Now().Truncate(time.Millisecond))),
 	}
 
-	res, err := c.Consent.Create(context.Background(), body)
+	res, err := c.RoleAssignment.Create(context.Background(), body)
 	assert.NoError(t, err)
 
 	t.Run("CheckCreate", func(t *testing.T) {
 		assert.NotEmpty(t, res.Data.UID)
 
-		assert.Equal(t, body.OrgUID, res.Data.OrgUID)
 		assert.Equal(t, body.PersonUID, res.Data.PersonUID)
-		assert.Equal(t, body.Purpose, res.Data.Purpose)
+		assert.Equal(t, body.RoleUID, res.Data.RoleUID)
 		assert.True(t, body.ValidFrom.Equal(*res.Data.ValidFrom))
 		assert.Equal(t, body.ValidTo, res.Data.ValidTo)
 
@@ -56,14 +53,13 @@ func TestCreateAndUpdateConsent(t *testing.T) {
 
 	t.Run("CheckUpdate", func(t *testing.T) {
 		body.ValidTo = lo.ToPtr(strfmt.DateTime(time.Now().Truncate(time.Millisecond)))
-		res, err := c.Consent.Update(context.Background(), *res.Data.UID, body)
+		res, err := c.RoleAssignment.Update(context.Background(), *res.Data.UID, body)
 
 		assert.NoError(t, err)
 		assert.NotEmpty(t, res.Data.UID)
 
-		assert.Equal(t, body.OrgUID, res.Data.OrgUID)
 		assert.Equal(t, body.PersonUID, res.Data.PersonUID)
-		assert.Equal(t, body.Purpose, res.Data.Purpose)
+		assert.Equal(t, body.RoleUID, res.Data.RoleUID)
 		assert.True(t, body.ValidFrom.Equal(*res.Data.ValidFrom))
 		assert.True(t, body.ValidTo.Equal(*res.Data.ValidTo))
